requirements.txt
adbutils==1.2.15
altgraph==0.17.4
apkutils2==1.0.0
av==9.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
cigam==0.0.3
coloredlogs==15.0.1
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
deprecation==2.1.0
flatbuffers==24.12.23
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.1
onnxruntime-gpu==1.20.1
opencv-python==4.10.0.84
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
protobuf==5.29.2
py==1.11.0
pyelftools==0.31
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyreadline3==3.5.4
pywin32-ctypes==0.2.3
requests==2.32.3
retry==0.9.2
scrcpy-client==0.4.7
soupsieve==2.6
sympy==1.13.3
urllib3==2.3.0
whichcraft==0.6.1
xmltodict==0.14.2
